pyproject.toml
[project]
name = "dagster-uc"
version = "0.1.0"
authors = [{name = "Stefan Verbruggen"}, {name = "Ion Koutsouris"}]
dependencies = [
    "kr8s < 1.0",
    "pyhelm3==0.3.3",
    "typer==0.12.3",
    "pyyaml",
    "pytz"
]

[build-system]
requires = ["setuptools", "wheel"]

[project.scripts]
dagster-uc = "dagster_uc.manage_user_code_deployments:app"


[tool.setuptools]
packages = ["dagster_uc"]

[tool.pyright]

typeCheckingMode = 'basic'
reportUnknownMemberType   = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
reportMissingImports = false
pythonVersion =  "3.10"

[tool.ruff]
force-exclude = true
lint.select = [
    "E",
    "F",
    "W",
    "N",
    "D",
    "UP",
    "ANN",
    "A",
    "COM",
    "C4",
    "ICN",
    "PT",
    "SIM",
    "ARG",
    "TD",
    "ERA",
    "NPY",
    "I",
]

lint.ignore = [
    "D100",
    "D104",
    "D107",
    "D203",
    "D205",
    "D206",
    "D213",
    "D400",
    "D401",
    "D413",
    "D415",
    "E501",
    "E712",
    "PT011",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "ANN204",
    "C416",
    "TD003",
]


# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    ".ipynb",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true

[tool.ruff.lint.per-file-ignores]
"dagster_uc_tests/*.py" = ["D100", "D103", "B018", "ANN001"]


[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"